k8sUtils creation, event metric for ns health
