+++ libraries/RW/K8s/__init__.py
@@ -1,1 +1,2 @@
 from .k8s import K8s
+from .k8sutils import K8sUtils
